@@ -17,12 +17,12 @@ typedef VoidFutureCallBack = Future<void> Function();
 
 typedef void OffsetCallBack(double offset);
 
-typedef void ModeChangeCallBack<T>(T mode);
+typedef void ModeChangeCallBack<T>(T? mode);
 
 /// a widget  implements ios pull down refresh effect and Android material RefreshIndicator overScroll effect
 abstract class RefreshIndicator extends StatefulWidget {
   /// refresh display style
-  final RefreshStyle refreshStyle;
+  final RefreshStyle? refreshStyle;
 
   /// the visual extent indicator
   final double height;
@@ -34,7 +34,7 @@ abstract class RefreshIndicator extends StatefulWidget {
   final Duration completeDuration;
 
   const RefreshIndicator(
-      {Key key,
+      {Key? key,
       this.height: 60.0,
       this.offset: 0.0,
       this.completeDuration: const Duration(milliseconds: 500),
@@ -51,10 +51,10 @@ abstract class LoadIndicator extends StatefulWidget {
   final double height;
 
   /// callback when user click footer
-  final VoidCallback onClick;
+  final VoidCallback? onClick;
 
   const LoadIndicator(
-      {Key key,
+      {Key? key,
       this.onClick,
       this.loadStyle: LoadStyle.ShowAlways,
       this.height: 60.0})
@@ -134,7 +134,7 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
     extends State<T>
     with IndicatorStateMixin<T, RefreshStatus>, RefreshProcessor {
   bool _inVisual() {
-    return _position.pixels < 0.0;
+    return _position!.pixels < 0.0;
   }
 
   double _calculateScrollOffset() {
@@ -142,10 +142,10 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
             ? (mode == RefreshStatus.twoLeveling ||
                     mode == RefreshStatus.twoLevelOpening ||
                     mode == RefreshStatus.twoLevelClosing
-                ? SmartRefresher.ofState(context).viewportExtent
+                ? SmartRefresher.ofState(context)!.viewportExtent
                 : widget.height)
             : 0.0) -
-        _position?.pixels;
+        (_position?.pixels as num);
   }
 
   @override
@@ -159,9 +159,9 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
   // handle the  state change between canRefresh and idle canRefresh  before refreshing
   void _dispatchModeByOffset(double offset) {
     if (mode == RefreshStatus.twoLeveling) {
-      if (_position.pixels > configuration.closeTwoLevelDistance &&
+      if (_position!.pixels > configuration!.closeTwoLevelDistance &&
           activity is BallisticScrollActivity) {
-        refresher.controller.twoLevelComplete();
+        refresher!.controller.twoLevelComplete();
         return;
       }
     }
@@ -176,19 +176,19 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
     }
 
     // If FrontStyle overScroll,it shouldn't disable gesture in scrollable
-    if (_position.extentBefore == 0.0 &&
+    if (_position!.extentBefore == 0.0 &&
         widget.refreshStyle == RefreshStyle.Front) {
-      _position.context.setIgnorePointer(false);
+      _position!.context.setIgnorePointer(false);
     }
     // Sometimes different devices return velocity differently, so it's impossible to judge from velocity whether the user
     // has invoked animateTo (0.0) or the user is dragging the view.Sometimes animateTo (0.0) does not return velocity = 0.0
     // velocity < 0.0 may be spring up,>0.0 spring down
-    if ((configuration.enableBallisticRefresh && activity.velocity < 0.0) ||
+    if ((configuration!.enableBallisticRefresh && activity!.velocity < 0.0) ||
         activity is DragScrollActivity ||
         activity is DrivenScrollActivity) {
-      if (refresher.enablePullDown &&
-          offset >= configuration.headerTriggerDistance) {
-        if (!configuration.skipCanRefresh) {
+      if (refresher!.enablePullDown &&
+          offset >= configuration!.headerTriggerDistance) {
+        if (!configuration!.skipCanRefresh) {
           mode = RefreshStatus.canRefresh;
         } else {
           floating = true;
@@ -198,13 +198,13 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
             mode = RefreshStatus.refreshing;
           });
         }
-      } else if (refresher.enablePullDown) {
+      } else if (refresher!.enablePullDown) {
         mode = RefreshStatus.idle;
       }
-      if (refresher.enableTwoLevel &&
-          offset >= configuration.twiceTriggerDistance) {
+      if (refresher!.enableTwoLevel &&
+          offset >= configuration!.twiceTriggerDistance) {
         mode = RefreshStatus.canTwoLevel;
-      } else if (refresher.enableTwoLevel && !refresher.enablePullDown) {
+      } else if (refresher!.enableTwoLevel && !refresher!.enablePullDown) {
         mode = RefreshStatus.idle;
       }
     }
@@ -241,15 +241,15 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
       resetValue();
 
       if (mode == RefreshStatus.idle)
-        SmartRefresher.ofState(context).setCanDrag(true);
+        SmartRefresher.ofState(context)!.setCanDrag(true);
     }
     if (mode == RefreshStatus.completed || mode == RefreshStatus.failed) {
       endRefresh().then((_) {
         if (!mounted) return;
         floating = false;
         if (mode == RefreshStatus.completed || mode == RefreshStatus.failed) {
-          SmartRefresher.ofState(context)
-              .setCanDrag(configuration.enableScrollWhenRefreshCompleted);
+          SmartRefresher.ofState(context)!
+              .setCanDrag(configuration!.enableScrollWhenRefreshCompleted);
         }
         update();
         /*
@@ -258,20 +258,20 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
           the _onOffsetChange didn't callback,it will keep failed or success state.
           2. As FrontStyle,when user dragging in 0~100 in refreshing state,it should be reset after the state change
           */
-        WidgetsBinding.instance.addPostFrameCallback((_) {
+        WidgetsBinding.instance!.addPostFrameCallback((_) {
           if (!mounted) {
             return;
           }
           if (widget.refreshStyle == RefreshStyle.Front) {
             if (_inVisual()) {
-              _position.jumpTo(0.0);
+              _position!.jumpTo(0.0);
             }
             mode = RefreshStatus.idle;
           } else {
             if (!_inVisual()) {
               mode = RefreshStatus.idle;
             } else {
-              activity.delegate.goBallistic(0.0);
+              activity!.delegate.goBallistic(0.0);
             }
           }
         });
@@ -281,32 +281,32 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
         floating = true;
         readyToRefresh();
       }
-      if (configuration.enableRefreshVibrate ?? false) {
+      if (configuration!.enableRefreshVibrate) {
         HapticFeedback.vibrate();
       }
-      if (refresher.onRefresh != null) refresher.onRefresh();
+      if (refresher!.onRefresh != null) refresher!.onRefresh!();
     } else if (mode == RefreshStatus.twoLevelOpening) {
       floating = true;
-      SmartRefresher.ofState(context).setCanDrag(false);
-      WidgetsBinding.instance.addPostFrameCallback((_) {
+      SmartRefresher.ofState(context)!.setCanDrag(false);
+      WidgetsBinding.instance!.addPostFrameCallback((_) {
         if (!mounted) return;
-        activity.resetActivity();
-        _position
+        activity!.resetActivity();
+        _position!
             .animateTo(0.0,
                 duration: const Duration(milliseconds: 500),
                 curve: Curves.linear)
             .whenComplete(() {
           mode = RefreshStatus.twoLeveling;
         });
-        if (refresher.onTwoLevel != null) refresher.onTwoLevel();
+        if (refresher!.onTwoLevel != null) refresher!.onTwoLevel!();
       });
     } else if (mode == RefreshStatus.twoLevelClosing) {
       floating = false;
-      SmartRefresher.ofState(context).setCanDrag(false);
+      SmartRefresher.ofState(context)!.setCanDrag(false);
       update();
     } else if (mode == RefreshStatus.twoLeveling) {
-      SmartRefresher.ofState(context)
-          .setCanDrag(configuration.enableScrollWhenTwoLevel);
+      SmartRefresher.ofState(context)!
+          .setCanDrag(configuration!.enableScrollWhenTwoLevel);
     }
     onModeChange(mode);
   }
@@ -334,7 +334,7 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
         child: RotatedBox(
           child: buildContent(context, mode),
           quarterTurns: needReverseAll() &&
-                  Scrollable.of(context).axisDirection == AxisDirection.up
+                  Scrollable.of(context)!.axisDirection == AxisDirection.up
               ? 10
               : 0,
         ),
@@ -342,7 +342,7 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
         refreshIndicatorLayoutExtent: mode == RefreshStatus.twoLeveling ||
                 mode == RefreshStatus.twoLevelOpening ||
                 mode == RefreshStatus.twoLevelClosing
-            ? SmartRefresher.ofState(context).viewportExtent
+            ? SmartRefresher.ofState(context)!.viewportExtent
             : widget.height,
         refreshStyle: widget.refreshStyle);
   }
@@ -353,11 +353,11 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
   // use to update between one page and above one page
   bool _isHide = false;
   bool _enableLoading = false;
-  LoadStatus _lastMode = LoadStatus.idle;
+  LoadStatus? _lastMode = LoadStatus.idle;
 
   double _calculateScrollOffset() {
     final double overScrollPastEnd =
-        math.max(_position.pixels - _position.maxScrollExtent, 0.0);
+        math.max(_position!.pixels - _position!.maxScrollExtent, 0.0);
     return overScrollPastEnd;
   }
 
@@ -390,10 +390,10 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
       }
 
       // this line for patch bug temporary:indicator disappears fastly when load more complete
-      if (mounted) Scrollable.of(context).position.correctBy(0.00001);
-      WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (mounted) Scrollable.of(context)!.position.correctBy(0.00001);
+      WidgetsBinding.instance!.addPostFrameCallback((_) {
         if (mounted && _position?.outOfRange == true) {
-          activity.delegate.goBallistic(0);
+          activity!.delegate.goBallistic(0);
         }
       });
       setState(() {
@@ -403,21 +403,21 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
   }
 
   bool _checkIfCanLoading() {
-    if (_position.maxScrollExtent - _position.pixels <=
-            configuration.footerTriggerDistance &&
-        _position.extentBefore > 2.0 &&
+    if (_position!.maxScrollExtent - _position!.pixels <=
+            configuration!.footerTriggerDistance &&
+        _position!.extentBefore > 2.0 &&
         _enableLoading) {
-      if (!configuration.autoLoad && mode == LoadStatus.idle) {
+      if (!configuration!.autoLoad && mode == LoadStatus.idle) {
         return false;
       }
-      if (!configuration.enableLoadingWhenFailed && mode == LoadStatus.failed) {
+      if (!configuration!.enableLoadingWhenFailed && mode == LoadStatus.failed) {
         return false;
       }
-      if (!configuration.enableLoadingWhenNoData && mode == LoadStatus.noMore) {
+      if (!configuration!.enableLoadingWhenNoData && mode == LoadStatus.noMore) {
         return false;
       }
       if (mode != LoadStatus.canLoading &&
-          _position.userScrollDirection == ScrollDirection.forward) {
+          _position!.userScrollDirection == ScrollDirection.forward) {
         return false;
       }
       return true;
@@ -436,11 +436,11 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
         mode == LoadStatus.noMore) {
       // #292,#265,#208
       // stop the slow bouncing when load more too fast
-      if (_position.activity.velocity < 0 &&
+      if (_position!.activity!.velocity < 0 &&
           _lastMode == LoadStatus.loading &&
-          !_position.outOfRange &&
+          !_position!.outOfRange &&
           _position is ScrollActivityDelegate) {
-        _position.beginActivity(
+        _position!.beginActivity(
             IdleScrollActivity(_position as ScrollActivityDelegate));
       }
 
@@ -450,11 +450,11 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
       if (!floating) {
         enterLoading();
       }
-      if (configuration.enableLoadMoreVibrate ?? false) {
+      if (configuration!.enableLoadMoreVibrate) {
         HapticFeedback.vibrate();
       }
-      if (refresher.onLoading != null) {
-        refresher.onLoading();
+      if (refresher!.onLoading != null) {
+        refresher!.onLoading!();
       }
       if (widget.loadStyle == LoadStyle.ShowWhenLoading) {
         floating = true;
@@ -478,7 +478,7 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
       }
     }
     if (activity is BallisticScrollActivity) {
-      if (configuration.enableBallisticLoad ?? true) {
+      if (configuration!.enableBallisticLoad) {
         if (_checkIfCanLoading()) enterLoading();
       } else if (mode == LoadStatus.canLoading) {
         enterLoading();
@@ -496,7 +496,7 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
   }
 
   void _listenScrollEnd() {
-    if (!_position.isScrollingNotifier.value) {
+    if (!_position!.isScrollingNotifier.value) {
       // when user release gesture from screen
       if (_isHide || mode == LoadStatus.loading || mode == LoadStatus.noMore) {
         return;
@@ -504,8 +504,8 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
 
       if (_checkIfCanLoading()) {
         if (activity is IdleScrollActivity) {
-          if ((configuration.enableBallisticLoad ?? true) ||
-              ((!configuration.enableBallisticLoad ?? true) &&
+          if ((configuration!.enableBallisticLoad) ||
+              ((!configuration!.enableBallisticLoad) &&
                   mode == LoadStatus.canLoading)) enterLoading();
         }
       }
@@ -517,8 +517,8 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
   }
 
   void _onPositionUpdated(ScrollPosition newPosition) {
-    _position?.isScrollingNotifier?.removeListener(_listenScrollEnd);
-    newPosition?.isScrollingNotifier?.addListener(_listenScrollEnd);
+    _position?.isScrollingNotifier.removeListener(_listenScrollEnd);
+    newPosition.isScrollingNotifier.addListener(_listenScrollEnd);
     super._onPositionUpdated(newPosition);
   }
 
@@ -532,7 +532,7 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
   @override
   void dispose() {
     // TODO: implement dispose
-    _position?.isScrollingNotifier?.removeListener(_listenScrollEnd);
+    _position?.isScrollingNotifier.removeListener(_listenScrollEnd);
     super.dispose();
   }
 
@@ -540,14 +540,14 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
   Widget build(BuildContext context) {
     // TODO: implement build
     return SliverLoading(
-        hideWhenNotFull: configuration.hideFooterWhenNotFull,
+        hideWhenNotFull: configuration!.hideFooterWhenNotFull,
         floating: widget.loadStyle == LoadStyle.ShowAlways
             ? true
             : widget.loadStyle == LoadStyle.HideAlways
                 ? false
                 : floating,
-        shouldFollowContent: configuration.shouldFooterFollowWhenNotFull != null
-            ? configuration.shouldFooterFollowWhenNotFull(mode)
+        shouldFollowContent: configuration!.shouldFooterFollowWhenNotFull != null
+            ? configuration!.shouldFooterFollowWhenNotFull!(mode)
             : mode == LoadStatus.noMore,
         layoutExtent: widget.height,
         mode: mode,
@@ -557,12 +557,12 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
             return GestureDetector(
               behavior: HitTestBehavior.opaque,
               onTap: () {
-                if ((mode == LoadStatus.idle && !configuration.autoLoad) ||
-                    (_mode.value == LoadStatus.failed)) {
+                if ((mode == LoadStatus.idle && !configuration!.autoLoad) ||
+                    (_mode!.value == LoadStatus.failed)) {
                   enterLoading();
                 }
                 if (widget.onClick != null) {
-                  widget.onClick();
+                  widget.onClick!();
                 }
               },
               child: buildContent(context, mode),
@@ -576,9 +576,9 @@ abstract class LoadIndicatorState<T extends LoadIndicator> extends State<T>
 ///
 /// help to finish the work that the header indicator and footer indicator need to do
 mixin IndicatorStateMixin<T extends StatefulWidget, V> on State<T> {
-  SmartRefresher refresher;
+  SmartRefresher? refresher;
 
-  RefreshConfiguration configuration;
+  RefreshConfiguration? configuration;
 
   bool _floating = false;
 
@@ -590,13 +590,13 @@ mixin IndicatorStateMixin<T extends StatefulWidget, V> on State<T> {
 
   get mode => _mode?.value;
 
-  ValueNotifier<V> _mode;
+  ValueNotifier<V?>? _mode;
 
-  ScrollActivity get activity => _position.activity;
+  ScrollActivity? get activity => _position!.activity;
 
   // it doesn't support get the ScrollController as the listener, because it will cause "multiple scrollview use one ScollController"
   // error,only replace the ScrollPosition to listen the offset
-  ScrollPosition _position;
+  ScrollPosition? _position;
 
   // update ui
   void update() {
@@ -611,8 +611,8 @@ mixin IndicatorStateMixin<T extends StatefulWidget, V> on State<T> {
     if (overscrollPast < 0.0) {
       return;
     }
-    if (refresher.onOffsetChange != null) {
-      refresher.onOffsetChange(V == RefreshStatus, overscrollPast);
+    if (refresher!.onOffsetChange != null) {
+      refresher!.onOffsetChange!(V == RefreshStatus, overscrollPast);
     }
     _dispatchModeByOffset(overscrollPast);
   }
@@ -627,10 +627,10 @@ mixin IndicatorStateMixin<T extends StatefulWidget, V> on State<T> {
   void _updateListener() {
     configuration = RefreshConfiguration.of(context);
     refresher = SmartRefresher.of(context);
-    ValueNotifier<V> newMode = V == RefreshStatus
-        ? refresher.controller.headerMode
-        : refresher.controller.footerMode;
-    final ScrollPosition newPosition = Scrollable.of(context).position;
+    ValueNotifier<V>? newMode = V == RefreshStatus
+        ? refresher!.controller.headerMode as ValueNotifier<V>?
+        : refresher!.controller.footerMode as ValueNotifier<V>?;
+    final ScrollPosition newPosition = Scrollable.of(context)!.position;
     if (newMode != _mode) {
       _mode?.removeListener(_handleModeChange);
       _mode = newMode;
@@ -648,7 +648,7 @@ mixin IndicatorStateMixin<T extends StatefulWidget, V> on State<T> {
   void initState() {
     // TODO: implement initState
     if (V == RefreshStatus) {
-      SmartRefresher.of(context)?.controller?.headerMode?.value =
+      SmartRefresher.of(context)?.controller.headerMode?.value =
           RefreshStatus.idle;
     }
     super.initState();
@@ -679,7 +679,7 @@ mixin IndicatorStateMixin<T extends StatefulWidget, V> on State<T> {
   }
 
   void _onPositionUpdated(ScrollPosition newPosition) {
-    refresher.controller.onPositionUpdated(newPosition);
+    refresher!.controller.onPositionUpdated(newPosition);
   }
 
   void _handleModeChange();
@@ -697,7 +697,7 @@ abstract class RefreshProcessor {
   void onOffsetChange(double offset) {}
 
   /// mode change callback
-  void onModeChange(RefreshStatus mode) {}
+  void onModeChange(RefreshStatus? mode) {}
 
   /// when indicator is ready into refresh,it will call back and waiting for this function finish,then callback onRefresh
   Future readyToRefresh() {
@@ -717,7 +717,7 @@ abstract class RefreshProcessor {
 abstract class LoadingProcessor {
   void onOffsetChange(double offset) {}
 
-  void onModeChange(LoadStatus mode) {}
+  void onModeChange(LoadStatus? mode) {}
 
   /// when indicator is ready into refresh,it will call back and waiting for this function finish,then callback onRefresh
   Future readyToLoad() {
