@@ -25,7 +25,7 @@ import 'indicator/material_indicator.dart';
 typedef void OnOffsetChange(bool up, double offset);
 
 /// when viewport not full one page, for different state,whether it should follow the content
-typedef bool ShouldFollowContent(LoadStatus status);
+typedef bool ShouldFollowContent(LoadStatus? status);
 
 /// global default indicator builder
 typedef IndicatorBuilder = Widget Function();
@@ -124,7 +124,7 @@ class SmartRefresher extends StatefulWidget {
   ///
   /// notice that: If child is  extends ScrollView,It will help you get the internal slivers and add footer and header in it.
   /// else it will put child into SliverToBoxAdapter and add footer and header
-  final Widget child;
+  final Widget? child;
 
   /// header indicator displace before content
   ///
@@ -134,7 +134,7 @@ class SmartRefresher extends StatefulWidget {
   ///
   /// from 1.5.2,it has been change RefreshIndicator to Widget,but remember only pass sliver widget,
   /// if you pass not a sliver,it will throw error
-  final Widget header;
+  final Widget? header;
 
   /// footer indicator display after content
   ///
@@ -144,7 +144,7 @@ class SmartRefresher extends StatefulWidget {
   ///
   /// from 1.5.2,it has been change LoadIndicator to Widget,but remember only pass sliver widget,
   //  if you pass not a sliver,it will throw error
-  final Widget footer;
+  final Widget? footer;
   // This bool will affect whether or not to have the function of drop-up load.
   final bool enablePullUp;
 
@@ -158,51 +158,51 @@ class SmartRefresher extends StatefulWidget {
   ///
   /// when the callback is happening,you should use [RefreshController]
   /// to end refreshing state,else it will keep refreshing state
-  final VoidCallback onRefresh;
+  final VoidCallback? onRefresh;
 
   /// callback when footer loading more data
   ///
   /// when the callback is happening,you should use [RefreshController]
   /// to end loading state,else it will keep loading state
-  final VoidCallback onLoading;
+  final VoidCallback? onLoading;
 
   /// callback when header ready to twoLevel
   ///
   /// If you want to close twoLevel,you should use [RefreshController.closeTwoLevel]
-  final VoidCallback onTwoLevel;
+  final VoidCallback? onTwoLevel;
 
   /// callback when the indicator scroll out of edge
-  final OnOffsetChange onOffsetChange;
+  final OnOffsetChange? onOffsetChange;
 
   /// Controll inner state
   final RefreshController controller;
 
   /// child content builder
-  final RefresherBuilder builder;
+  final RefresherBuilder? builder;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final Axis scrollDirection;
+  final Axis? scrollDirection;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final bool reverse;
+  final bool? reverse;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final ScrollController scrollController;
+  final ScrollController? scrollController;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final bool primary;
+  final bool? primary;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final double cacheExtent;
+  final double? cacheExtent;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final int semanticChildCount;
+  final int? semanticChildCount;
 
   /// copy from ScrollView,for setting in SingleChildView,not ScrollView
-  final DragStartBehavior dragStartBehavior;
+  final DragStartBehavior? dragStartBehavior;
 
   /// creates a widget help attach the refresh and load more function
   /// controller must not be null,
@@ -213,8 +213,8 @@ class SmartRefresher extends StatefulWidget {
   /// If you don't need pull down refresh ,just enablePullDown = false,
   /// If you  need pull up load ,just enablePullUp = true
   SmartRefresher(
-      {Key key,
-      @required this.controller,
+      {Key? key,
+      required this.controller,
       this.child,
       this.header,
       this.footer,
@@ -233,8 +233,7 @@ class SmartRefresher extends StatefulWidget {
       this.physics,
       this.scrollDirection,
       this.scrollController})
-      : assert(controller != null),
-        builder = null,
+      : builder = null,
         super(key: key);
 
   /// creates a widget help attach the refresh and load more function
@@ -245,9 +244,9 @@ class SmartRefresher extends StatefulWidget {
   /// refresh above NestedScrollVIew,we must use this constrctor to implements refresh above NestedScrollView,but for now,NestedScrollView
   /// can not support overscroll out of edge
   SmartRefresher.builder(
-      {Key key,
-      @required this.controller,
-      @required this.builder,
+      {Key? key,
+      required this.controller,
+      required this.builder,
       this.enablePullDown: true,
       this.enablePullUp: false,
       this.enableTwoLevel: false,
@@ -255,8 +254,7 @@ class SmartRefresher extends StatefulWidget {
       this.onLoading,
       this.onTwoLevel,
       this.onOffsetChange})
-      : assert(controller != null),
-        header = null,
+      : header = null,
         footer = null,
         child = null,
         scrollController = null,
@@ -269,12 +267,12 @@ class SmartRefresher extends StatefulWidget {
         primary = null,
         super(key: key);
 
-  static SmartRefresher of(BuildContext context) {
-    return context?.findAncestorWidgetOfExactType<SmartRefresher>();
+  static SmartRefresher? of(BuildContext context) {
+    return context.findAncestorWidgetOfExactType<SmartRefresher>();
   }
 
-  static SmartRefresherState ofState(BuildContext context) {
-    return context?.findAncestorStateOfType<SmartRefresherState>();
+  static SmartRefresherState? ofState(BuildContext context) {
+    return context.findAncestorStateOfType<SmartRefresherState>();
   }
 
   @override
@@ -285,7 +283,7 @@ class SmartRefresher extends StatefulWidget {
 }
 
 class SmartRefresherState extends State<SmartRefresher> {
-  RefreshPhysics _physics;
+  RefreshPhysics? _physics;
   bool _updatePhysics = false;
   double viewportExtent = 0;
   bool _canDrag = true;
@@ -298,15 +296,15 @@ class SmartRefresherState extends State<SmartRefresher> {
   final LoadIndicator defaultFooter = ClassicFooter();
 
   //build slivers from child Widget
-  List<Widget> _buildSliversByChild(
-      BuildContext context, Widget child, RefreshConfiguration configuration) {
-    List<Widget> slivers;
+  List<Widget>? _buildSliversByChild(
+      BuildContext context, Widget? child, RefreshConfiguration? configuration) {
+    List<Widget>? slivers;
     if (child is ScrollView) {
       if (child is BoxScrollView) {
         //avoid system inject padding when own indicator top or bottom
         Widget sliver = child.buildChildLayout(context);
         if (child.padding != null) {
-          slivers = [SliverPadding(sliver: sliver, padding: child.padding)];
+          slivers = [SliverPadding(sliver: sliver, padding: child.padding!)];
         } else {
           slivers = [sliver];
         }
@@ -325,7 +323,7 @@ class SmartRefresherState extends State<SmartRefresher> {
           0,
           widget.header ??
               (configuration?.headerBuilder != null
-                  ? configuration?.headerBuilder()
+                  ? configuration?.headerBuilder!()
                   : null) ??
               defaultHeader);
     }
@@ -333,7 +331,7 @@ class SmartRefresherState extends State<SmartRefresher> {
     if (widget.enablePullUp) {
       slivers?.add(widget.footer ??
           (configuration?.footerBuilder != null
-              ? configuration?.footerBuilder()
+              ? configuration?.footerBuilder!()
               : null) ??
           defaultFooter);
     }
@@ -342,11 +340,11 @@ class SmartRefresherState extends State<SmartRefresher> {
   }
 
   ScrollPhysics _getScrollPhysics(
-      RefreshConfiguration conf, ScrollPhysics physics) {
+      RefreshConfiguration? conf, ScrollPhysics physics) {
     final bool isBouncingPhysics = physics is BouncingScrollPhysics ||
         (physics is AlwaysScrollableScrollPhysics &&
             ScrollConfiguration.of(context)
-                    ?.getScrollPhysics(context)
+                    .getScrollPhysics(context)
                     .runtimeType ==
                 BouncingScrollPhysics);
     return _physics = RefreshPhysics(
@@ -376,25 +374,25 @@ class SmartRefresherState extends State<SmartRefresher> {
   }
 
   // build the customScrollView
-  Widget _buildBodyBySlivers(
-      Widget childView, List<Widget> slivers, RefreshConfiguration conf) {
-    Widget body;
+  Widget? _buildBodyBySlivers(
+      Widget? childView, List<Widget>? slivers, RefreshConfiguration? conf) {
+    Widget? body;
     if (childView is! Scrollable) {
-      bool primary = widget.primary;
-      Key key;
-      double cacheExtent = widget.cacheExtent;
+      bool? primary = widget.primary;
+      Key? key;
+      double? cacheExtent = widget.cacheExtent;
 
-      Axis scrollDirection = widget.scrollDirection;
-      int semanticChildCount = widget.semanticChildCount;
-      bool reverse = widget.reverse;
-      ScrollController scrollController = widget.scrollController;
-      DragStartBehavior dragStartBehavior = widget.dragStartBehavior;
-      ScrollPhysics physics = widget.physics;
-      Key center;
-      double anchor;
-      ScrollViewKeyboardDismissBehavior keyboardDismissBehavior;
-      String restorationId;
-      Clip clipBehavior;
+      Axis? scrollDirection = widget.scrollDirection;
+      int? semanticChildCount = widget.semanticChildCount;
+      bool? reverse = widget.reverse;
+      ScrollController? scrollController = widget.scrollController;
+      DragStartBehavior? dragStartBehavior = widget.dragStartBehavior;
+      ScrollPhysics? physics = widget.physics;
+      Key? center;
+      double? anchor;
+      ScrollViewKeyboardDismissBehavior? keyboardDismissBehavior;
+      String? restorationId;
+      Clip? clipBehavior;
 
       if (childView is ScrollView) {
         primary = primary ?? childView.primary;
@@ -438,7 +436,7 @@ class SmartRefresherState extends State<SmartRefresher> {
         center: center,
         physics:
             _getScrollPhysics(conf, physics ?? AlwaysScrollableScrollPhysics()),
-        slivers: slivers,
+        slivers: slivers!,
         dragStartBehavior: dragStartBehavior ?? DragStartBehavior.start,
         reverse: reverse ?? false,
       );
@@ -451,20 +449,20 @@ class SmartRefresherState extends State<SmartRefresher> {
         semanticChildCount: childView.semanticChildCount,
         dragStartBehavior: childView.dragStartBehavior,
         viewportBuilder: (context, offset) {
-          Viewport viewport = childView.viewportBuilder(context, offset);
+          Viewport viewport = childView.viewportBuilder(context, offset) as Viewport;
           if (widget.enablePullDown) {
             viewport.children.insert(
                 0,
                 widget.header ??
                     (conf?.headerBuilder != null
-                        ? conf?.headerBuilder()
+                        ? conf?.headerBuilder!()
                         : null) ??
                     defaultHeader);
           }
           //insert header or footer
           if (widget.enablePullUp) {
             viewport.children.add(widget.footer ??
-                (conf?.footerBuilder != null ? conf?.footerBuilder() : null) ??
+                (conf?.footerBuilder != null ? conf?.footerBuilder!() : null) ??
                 defaultFooter);
           }
           return viewport;
@@ -475,18 +473,18 @@ class SmartRefresherState extends State<SmartRefresher> {
   }
 
   bool _ifNeedUpdatePhysics() {
-    RefreshConfiguration conf = RefreshConfiguration.of(context);
+    RefreshConfiguration? conf = RefreshConfiguration.of(context);
     if (conf == null || _physics == null) {
       return false;
     }
 
-    if (conf.topHitBoundary != _physics.topHitBoundary ||
-        _physics.bottomHitBoundary != conf.bottomHitBoundary ||
-        conf.maxOverScrollExtent != _physics.maxOverScrollExtent ||
-        _physics.maxUnderScrollExtent != conf.maxUnderScrollExtent ||
-        _physics.dragSpeedRatio != conf.dragSpeedRatio ||
-        _physics.enableScrollWhenTwoLevel != conf.enableScrollWhenTwoLevel ||
-        _physics.enableScrollWhenRefreshCompleted !=
+    if (conf.topHitBoundary != _physics!.topHitBoundary ||
+        _physics!.bottomHitBoundary != conf.bottomHitBoundary ||
+        conf.maxOverScrollExtent != _physics!.maxOverScrollExtent ||
+        _physics!.maxUnderScrollExtent != conf.maxUnderScrollExtent ||
+        _physics!.dragSpeedRatio != conf.dragSpeedRatio ||
+        _physics!.enableScrollWhenTwoLevel != conf.enableScrollWhenTwoLevel ||
+        _physics!.enableScrollWhenRefreshCompleted !=
             conf.enableScrollWhenRefreshCompleted) {
       return true;
     }
@@ -506,10 +504,10 @@ class SmartRefresherState extends State<SmartRefresher> {
   void didUpdateWidget(SmartRefresher oldWidget) {
     // TODO: implement didUpdateWidget
     if (widget.controller != oldWidget.controller) {
-      widget.controller.headerMode.value =
-          oldWidget.controller.headerMode.value;
-      widget.controller.footerMode.value =
-          oldWidget.controller.footerMode.value;
+      widget.controller.headerMode!.value =
+          oldWidget.controller.headerMode!.value;
+      widget.controller.footerMode!.value =
+          oldWidget.controller.footerMode!.value;
     }
     super.didUpdateWidget(oldWidget);
   }
@@ -527,7 +525,7 @@ class SmartRefresherState extends State<SmartRefresher> {
   void initState() {
     // TODO: implement initState
     if (widget.controller.initialRefresh) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
+      WidgetsBinding.instance!.addPostFrameCallback((_) {
         //  if mounted,it avoid one situation: when init done,then dispose the widget before build.
         //  this   situation mostly TabBarView
         if (mounted) widget.controller.requestRefresh();
@@ -545,23 +543,23 @@ class SmartRefresherState extends State<SmartRefresher> {
 
   @override
   Widget build(BuildContext context) {
-    final RefreshConfiguration configuration = RefreshConfiguration.of(context);
-    Widget body;
+    final RefreshConfiguration? configuration = RefreshConfiguration.of(context);
+    Widget? body;
     if (widget.builder != null)
-      body = widget.builder(context,
-          _getScrollPhysics(configuration, AlwaysScrollableScrollPhysics()));
+      body = widget.builder!(context,
+          _getScrollPhysics(configuration, AlwaysScrollableScrollPhysics()) as RefreshPhysics);
     else {
-      List<Widget> slivers =
+      List<Widget>? slivers =
           _buildSliversByChild(context, widget.child, configuration);
       body = _buildBodyBySlivers(widget.child, slivers, configuration);
     }
     if (configuration == null) {
-      body = RefreshConfiguration(child: body);
+      body = RefreshConfiguration(child: body!);
     }
     return LayoutBuilder(
       builder: (c2, cons) {
         viewportExtent = cons.biggest.height;
-        return body;
+        return body!;
       },
     );
   }
@@ -575,25 +573,25 @@ class SmartRefresherState extends State<SmartRefresher> {
 /// * [SmartRefresher],a widget help you attach refresh and load more function easily
 class RefreshController {
   /// header status mode controll
-  ValueNotifier<RefreshStatus> headerMode;
+  ValueNotifier<RefreshStatus>? headerMode;
 
   /// footer status mode controll
-  ValueNotifier<LoadStatus> footerMode;
+  ValueNotifier<LoadStatus>? footerMode;
 
   /// the scrollable inner's position
   ///
   /// notice that: position is null before build,
   /// the value is get when the header or footer callback onPositionUpdated
-  ScrollPosition position;
+  ScrollPosition? position;
 
   /// deprecated member,not suggest to use it,it contain share position bug
   @Deprecated(
       'advice set ScrollController to child,use it directly will cause bug when call jumpTo() and animateTo()')
-  ScrollController scrollController;
+  ScrollController? scrollController;
 
-  RefreshStatus get headerStatus => headerMode?.value;
+  RefreshStatus? get headerStatus => headerMode?.value;
 
-  LoadStatus get footerStatus => footerMode?.value;
+  LoadStatus? get footerStatus => footerMode?.value;
 
   bool get isRefresh => headerMode?.value == RefreshStatus.refreshing;
 
@@ -613,37 +611,33 @@ class RefreshController {
   /// initialLoadStatus: footerMode default value
   RefreshController(
       {this.initialRefresh: false,
-      RefreshStatus initialRefreshStatus,
-      LoadStatus initialLoadStatus}) {
+      RefreshStatus? initialRefreshStatus,
+      LoadStatus? initialLoadStatus}) {
     this.headerMode = ValueNotifier(initialRefreshStatus ?? RefreshStatus.idle);
     this.footerMode = ValueNotifier(initialLoadStatus ?? LoadStatus.idle);
   }
 
   /// callback when the indicator is builded,and catch the scrollable's inner position
   void onPositionUpdated(ScrollPosition newPosition) {
-    assert(newPosition != null);
-    position?.isScrollingNotifier?.removeListener(_listenScrollEnd);
+    position?.isScrollingNotifier.removeListener(_listenScrollEnd);
     position = newPosition;
-    position.isScrollingNotifier.addListener(_listenScrollEnd);
+    position!.isScrollingNotifier.addListener(_listenScrollEnd);
   }
 
   void _detachPosition() {
-    position?.isScrollingNotifier?.removeListener(_listenScrollEnd);
+    position?.isScrollingNotifier.removeListener(_listenScrollEnd);
   }
 
-  StatefulElement _findIndicator(BuildContext context, Type elementType) {
-    if (context == null) {
-      return null;
-    }
-    StatefulElement result;
+  StatefulElement? _findIndicator(BuildContext context, Type elementType) {
+    StatefulElement? result;
     context.visitChildElements((Element e) {
       if (elementType == RefreshIndicator) {
         if (e.widget is RefreshIndicator) {
-          result = e;
+          result = e as StatefulElement?;
         }
       } else {
         if (e.widget is LoadIndicator) {
-          result = e;
+          result = e as StatefulElement?;
         }
       }
 
@@ -656,42 +650,42 @@ class RefreshController {
   /// but ScrollPhysics didn't provide one way to spring back when outOfEdge(stopped by applyBouncingCondition return != 0.0)
   /// so for making it spring back, it should be trigger goBallistic make it spring back
   void _listenScrollEnd() {
-    if (position != null && position.outOfRange) {
+    if (position != null && position!.outOfRange) {
       position?.activity?.applyNewDimensions();
     }
   }
 
   /// make the header enter refreshing state,and callback onRefresh
-  Future<void> requestRefresh(
+  Future<void>? requestRefresh(
       {bool needMove: true,
       Duration duration: const Duration(milliseconds: 500),
       Curve curve: Curves.linear}) {
     assert(position != null,
         'Try not to call requestRefresh() before build,please call after the ui was rendered');
     if (isRefresh) return Future.value();
-    StatefulElement indicatorElement =
-        _findIndicator(position.context.storageContext, RefreshIndicator);
+    StatefulElement? indicatorElement =
+        _findIndicator(position!.context.storageContext, RefreshIndicator);
     if (indicatorElement == null) return null;
-    (indicatorElement.state as RefreshIndicatorState)?.floating = true;
+    (indicatorElement.state as RefreshIndicatorState).floating = true;
     if (needMove)
-      SmartRefresher.ofState(position.context.storageContext)
+      SmartRefresher.ofState(position!.context.storageContext)
           ?.setCanDrag(false);
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       if (needMove) {
-        return Future.delayed(const Duration(milliseconds: 50)).then((_) async {
+        Future.delayed(const Duration(milliseconds: 50)).then((_) async {
           // - 0.0001 is for NestedScrollView.
           await position
-              ?.animateTo(position.minScrollExtent - 0.0001,
+              ?.animateTo(position!.minScrollExtent - 0.0001,
                   duration: duration, curve: curve)
-              ?.then((_) {
-            SmartRefresher.ofState(position.context.storageContext)
+              .then((_) {
+            SmartRefresher.ofState(position!.context.storageContext)
                 ?.setCanDrag(true);
-            headerMode.value = RefreshStatus.refreshing;
+            headerMode!.value = RefreshStatus.refreshing;
           });
         });
       } else {
-        return Future.value().then((_) {
-          headerMode.value = RefreshStatus.refreshing;
+        Future.value().then((_) {
+          headerMode!.value = RefreshStatus.refreshing;
         });
       }
     });
@@ -704,42 +698,42 @@ class RefreshController {
       Curve curve: Curves.linear}) {
     assert(position != null,
         'Try not to call requestRefresh() before build,please call after the ui was rendered');
-    headerMode.value = RefreshStatus.twoLevelOpening;
+    headerMode!.value = RefreshStatus.twoLevelOpening;
     return Future.delayed(const Duration(milliseconds: 50)).then((_) async {
-      await position?.animateTo(position.minScrollExtent,
+      await position?.animateTo(position!.minScrollExtent,
           duration: duration, curve: curve);
     });
   }
 
   /// make the footer enter loading state,and callback onLoading
-  Future<void> requestLoading(
+  Future<void>? requestLoading(
       {bool needMove: true,
       Duration duration: const Duration(milliseconds: 300),
       Curve curve: Curves.linear}) {
     assert(position != null,
         'Try not to call requestLoading() before build,please call after the ui was rendered');
     if (isLoading) return Future.value();
-    StatefulElement indicatorElement =
-        _findIndicator(position.context.storageContext, LoadIndicator);
+    StatefulElement? indicatorElement =
+        _findIndicator(position!.context.storageContext, LoadIndicator);
     if (indicatorElement == null) return null;
-    (indicatorElement.state as LoadIndicatorState)?.floating = true;
+    (indicatorElement.state as LoadIndicatorState).floating = true;
     if (needMove)
-      SmartRefresher.ofState(position.context.storageContext)
+      SmartRefresher.ofState(position!.context.storageContext)
           ?.setCanDrag(false);
     if (needMove) {
       return Future.delayed(const Duration(milliseconds: 50)).then((_) async {
         await position
-            ?.animateTo(position.maxScrollExtent,
+            ?.animateTo(position!.maxScrollExtent,
                 duration: duration, curve: curve)
-            ?.then((_) {
-          SmartRefresher.ofState(position.context.storageContext)
+            .then((_) {
+          SmartRefresher.ofState(position!.context.storageContext)
               ?.setCanDrag(true);
-          footerMode.value = LoadStatus.loading;
+          footerMode!.value = LoadStatus.loading;
         });
       });
     } else {
       return Future.value().then((_) {
-        footerMode.value = LoadStatus.loading;
+        footerMode!.value = LoadStatus.loading;
       });
     }
   }
@@ -755,15 +749,15 @@ class RefreshController {
   }
 
   /// end twoLeveling,will return back first floor
-  Future<void> twoLevelComplete(
+  Future<void>? twoLevelComplete(
       {Duration duration: const Duration(milliseconds: 500),
       Curve curve: Curves.linear}) {
     headerMode?.value = RefreshStatus.twoLevelClosing;
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      return position
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
+      position!
           .animateTo(0.0, duration: duration, curve: curve)
           .whenComplete(() {
-        headerMode.value = RefreshStatus.idle;
+        headerMode!.value = RefreshStatus.idle;
       });
     });
     return null;
@@ -782,7 +776,7 @@ class RefreshController {
   /// after data returned,set the footer state to idle
   void loadComplete() {
     // change state after ui update,else it will have a bug:twice loading
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       footerMode?.value = LoadStatus.idle;
     });
   }
@@ -790,14 +784,14 @@ class RefreshController {
   /// If catchError happen,you may call loadFailed indicate fetch data from network failed
   void loadFailed() {
     // change state after ui update,else it will have a bug:twice loading
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       footerMode?.value = LoadStatus.failed;
     });
   }
 
   /// load more success without error,but no data returned
   void loadNoData() {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       footerMode?.value = LoadStatus.noMore;
     });
   }
@@ -805,14 +799,14 @@ class RefreshController {
   /// reset footer noData state  to idle
   void resetNoData() {
     if (footerMode?.value == LoadStatus.noMore) {
-      footerMode.value = LoadStatus.idle;
+      footerMode!.value = LoadStatus.idle;
     }
   }
 
   /// for some special situation, you should call dispose() for safe,it may throw errors after parent widget dispose
   void dispose() {
-    headerMode.dispose();
-    footerMode.dispose();
+    headerMode!.dispose();
+    footerMode!.dispose();
     headerMode = null;
     footerMode = null;
   }
@@ -829,10 +823,10 @@ class RefreshConfiguration extends InheritedWidget {
   final Widget child;
 
   /// global default header builder
-  final IndicatorBuilder headerBuilder;
+  final IndicatorBuilder? headerBuilder;
 
   /// global default footer builder
-  final IndicatorBuilder footerBuilder;
+  final IndicatorBuilder? footerBuilder;
 
   /// custom spring animate
   final SpringDescription springDescription;
@@ -841,7 +835,7 @@ class RefreshConfiguration extends InheritedWidget {
   final bool skipCanRefresh;
 
   /// if it should follow content for different state
-  final ShouldFollowContent shouldFooterFollowWhenNotFull;
+  final ShouldFollowContent? shouldFooterFollowWhenNotFull;
 
   /// when listView data small(not enough one page) , it should be hide
   final bool hideFooterWhenNotFull;
@@ -883,16 +877,16 @@ class RefreshConfiguration extends InheritedWidget {
   final double dragSpeedRatio;
 
   /// max overScroll distance when out of edge
-  final double maxOverScrollExtent;
+  final double? maxOverScrollExtent;
 
   /// 	max underScroll distance when out of edge
-  final double maxUnderScrollExtent;
+  final double? maxUnderScrollExtent;
 
   /// The boundary is located at the top edge and stops when inertia rolls over the boundary distance
-  final double topHitBoundary;
+  final double? topHitBoundary;
 
   /// The boundary is located at the bottom edge and stops when inertia rolls under the boundary distance
-  final double bottomHitBoundary;
+  final double? bottomHitBoundary;
 
   /// toggle of  refresh vibrate
   final bool enableRefreshVibrate;
@@ -901,7 +895,8 @@ class RefreshConfiguration extends InheritedWidget {
   final bool enableLoadMoreVibrate;
 
   RefreshConfiguration(
-      {@required this.child,
+      {Key? key,
+      required this.child,
       this.headerBuilder,
       this.footerBuilder,
       this.dragSpeedRatio: 1.0,
@@ -930,95 +925,96 @@ class RefreshConfiguration extends InheritedWidget {
       this.enableLoadMoreVibrate: false,
       this.topHitBoundary,
       this.bottomHitBoundary})
-      : assert(child != null),
-        assert(headerTriggerDistance > 0),
+      : assert(headerTriggerDistance > 0),
         assert(twiceTriggerDistance > 0),
         assert(closeTwoLevelDistance > 0),
-        assert(dragSpeedRatio > 0);
+        assert(dragSpeedRatio > 0),
+        super(key: key, child: child);
 
   /// Construct RefreshConfiguration to copy attributes from ancestor nodes
   /// If the parameter is null, it will automatically help you to absorb the attributes of your ancestor Refresh Configuration, instead of having to copy them manually by yourself.
   ///
   /// it mostly use in some stiuation is different the other SmartRefresher in App
   RefreshConfiguration.copyAncestor({
-    @required BuildContext context,
-    @required this.child,
-    IndicatorBuilder headerBuilder,
-    IndicatorBuilder footerBuilder,
-    double dragSpeedRatio,
-    ShouldFollowContent shouldFooterFollowWhenNotFull,
-    bool enableScrollWhenTwoLevel,
-    bool enableBallisticRefresh,
-    bool enableBallisticLoad,
-    bool enableLoadingWhenNoData,
-    SpringDescription springDescription,
-    bool enableScrollWhenRefreshCompleted,
-    bool enableLoadingWhenFailed,
-    double twiceTriggerDistance,
-    double closeTwoLevelDistance,
-    bool skipCanRefresh,
-    bool autoLoad,
-    double maxOverScrollExtent,
-    double maxUnderScrollExtent,
-    double topHitBoundary,
-    double bottomHitBoundary,
-    double headerTriggerDistance,
-    double footerTriggerDistance,
-    bool enableRefreshVibrate,
-    bool enableLoadMoreVibrate,
-    bool hideFooterWhenNotFull,
-  })  : assert(context != null, child != null),
-        assert(RefreshConfiguration.of(context) != null,
+    Key? key,
+    required BuildContext context,
+    required this.child,
+    IndicatorBuilder? headerBuilder,
+    IndicatorBuilder? footerBuilder,
+    double? dragSpeedRatio,
+    ShouldFollowContent? shouldFooterFollowWhenNotFull,
+    bool? enableScrollWhenTwoLevel,
+    bool? enableBallisticRefresh,
+    bool? enableBallisticLoad,
+    bool? enableLoadingWhenNoData,
+    SpringDescription? springDescription,
+    bool? enableScrollWhenRefreshCompleted,
+    bool? enableLoadingWhenFailed,
+    double? twiceTriggerDistance,
+    double? closeTwoLevelDistance,
+    bool? skipCanRefresh,
+    bool? autoLoad,
+    double? maxOverScrollExtent,
+    double? maxUnderScrollExtent,
+    double? topHitBoundary,
+    double? bottomHitBoundary,
+    double? headerTriggerDistance,
+    double? footerTriggerDistance,
+    bool? enableRefreshVibrate,
+    bool? enableLoadMoreVibrate,
+    bool? hideFooterWhenNotFull,
+  })  : assert(RefreshConfiguration.of(context) != null,
             "search RefreshConfiguration anscestor return null,please  Make sure that RefreshConfiguration is the ancestor of that element"),
-        autoLoad = autoLoad ?? RefreshConfiguration.of(context).autoLoad,
+        autoLoad = autoLoad ?? RefreshConfiguration.of(context)!.autoLoad,
         headerBuilder =
-            headerBuilder ?? RefreshConfiguration.of(context).headerBuilder,
+            headerBuilder ?? RefreshConfiguration.of(context)!.headerBuilder,
         footerBuilder =
-            footerBuilder ?? RefreshConfiguration.of(context).footerBuilder,
+            footerBuilder ?? RefreshConfiguration.of(context)!.footerBuilder,
         dragSpeedRatio =
-            dragSpeedRatio ?? RefreshConfiguration.of(context).dragSpeedRatio,
+            dragSpeedRatio ?? RefreshConfiguration.of(context)!.dragSpeedRatio,
         twiceTriggerDistance = twiceTriggerDistance ??
-            RefreshConfiguration.of(context).twiceTriggerDistance,
+            RefreshConfiguration.of(context)!.twiceTriggerDistance,
         headerTriggerDistance = headerTriggerDistance ??
-            RefreshConfiguration.of(context).headerTriggerDistance,
+            RefreshConfiguration.of(context)!.headerTriggerDistance,
         footerTriggerDistance = footerTriggerDistance ??
-            RefreshConfiguration.of(context).footerTriggerDistance,
+            RefreshConfiguration.of(context)!.footerTriggerDistance,
         springDescription = springDescription ??
-            RefreshConfiguration.of(context).springDescription,
+            RefreshConfiguration.of(context)!.springDescription,
         hideFooterWhenNotFull = hideFooterWhenNotFull ??
-            RefreshConfiguration.of(context).hideFooterWhenNotFull,
+            RefreshConfiguration.of(context)!.hideFooterWhenNotFull,
         maxOverScrollExtent = maxOverScrollExtent ??
-            RefreshConfiguration.of(context).maxOverScrollExtent,
+            RefreshConfiguration.of(context)!.maxOverScrollExtent,
         maxUnderScrollExtent = maxUnderScrollExtent ??
-            RefreshConfiguration.of(context).maxUnderScrollExtent,
+            RefreshConfiguration.of(context)!.maxUnderScrollExtent,
         topHitBoundary =
-            topHitBoundary ?? RefreshConfiguration.of(context).topHitBoundary,
+            topHitBoundary ?? RefreshConfiguration.of(context)!.topHitBoundary,
         bottomHitBoundary = bottomHitBoundary ??
-            RefreshConfiguration.of(context).bottomHitBoundary,
+            RefreshConfiguration.of(context)!.bottomHitBoundary,
         skipCanRefresh =
-            skipCanRefresh ?? RefreshConfiguration.of(context).skipCanRefresh,
+            skipCanRefresh ?? RefreshConfiguration.of(context)!.skipCanRefresh,
         enableScrollWhenRefreshCompleted = enableScrollWhenRefreshCompleted ??
-            RefreshConfiguration.of(context).enableScrollWhenRefreshCompleted,
+            RefreshConfiguration.of(context)!.enableScrollWhenRefreshCompleted,
         enableScrollWhenTwoLevel = enableScrollWhenTwoLevel ??
-            RefreshConfiguration.of(context).enableScrollWhenTwoLevel,
+            RefreshConfiguration.of(context)!.enableScrollWhenTwoLevel,
         enableBallisticRefresh = enableBallisticRefresh ??
-            RefreshConfiguration.of(context).enableBallisticRefresh,
+            RefreshConfiguration.of(context)!.enableBallisticRefresh,
         enableBallisticLoad = enableBallisticLoad ??
-            RefreshConfiguration.of(context).enableBallisticLoad,
+            RefreshConfiguration.of(context)!.enableBallisticLoad,
         enableLoadingWhenNoData = enableLoadingWhenNoData ??
-            RefreshConfiguration.of(context).enableLoadingWhenNoData,
+            RefreshConfiguration.of(context)!.enableLoadingWhenNoData,
         enableLoadingWhenFailed = enableLoadingWhenFailed ??
-            RefreshConfiguration.of(context).enableLoadingWhenFailed,
+            RefreshConfiguration.of(context)!.enableLoadingWhenFailed,
         closeTwoLevelDistance = closeTwoLevelDistance ??
-            RefreshConfiguration.of(context).closeTwoLevelDistance,
+            RefreshConfiguration.of(context)!.closeTwoLevelDistance,
         enableRefreshVibrate = enableRefreshVibrate ??
-            RefreshConfiguration.of(context).enableRefreshVibrate,
+            RefreshConfiguration.of(context)!.enableRefreshVibrate,
         enableLoadMoreVibrate = enableLoadMoreVibrate ??
-            RefreshConfiguration.of(context).enableLoadMoreVibrate,
+            RefreshConfiguration.of(context)!.enableLoadMoreVibrate,
         shouldFooterFollowWhenNotFull = shouldFooterFollowWhenNotFull ??
-            RefreshConfiguration.of(context).shouldFooterFollowWhenNotFull;
+            RefreshConfiguration.of(context)!.shouldFooterFollowWhenNotFull,
+        super(key: key, child: child);
 
-  static RefreshConfiguration of(BuildContext context) {
+  static RefreshConfiguration? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<RefreshConfiguration>();
   }
 
